requirements.txt
celery==5.2.1
Flask==2.0.1
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.10
Flask-Mail==0.9.1
Flask-SQLAlchemy==2.5.1
marshmallow==3.19.0
psycopg2-binary==2.9.1
python-dateutil==2.7.3
python-dotenv==1.0.0
redis==4.0.2
SQLAlchemy==1.3.8
tzdata==2021.4
tzlocal==4.0.1
Unidecode==1.3.6
virtualenv==20.7.2
